proto: move OpenSSL public key wrangling code into openssl module
