+++ crates/proto/src/dnssec/public_key.rs
@@ -6,27 +6,15 @@
 // copied, modified, or distributed except according to those terms.
 
 //! Public Key implementations for supported key types
-#[cfg(not(any(feature = "dnssec-openssl", feature = "dnssec-ring")))]
-use std::marker::PhantomData;
-use std::sync::Arc;
 
-#[cfg(feature = "dnssec-openssl")]
-use openssl::bn::BigNumContext;
-#[cfg(feature = "dnssec-openssl")]
-use openssl::ec::{EcKey, PointConversionForm};
-#[cfg(feature = "dnssec-openssl")]
-use openssl::nid::Nid;
-#[cfg(feature = "dnssec-openssl")]
-use openssl::pkey::HasPublic;
-#[cfg(feature = "dnssec-openssl")]
-use openssl::rsa::Rsa as OpenSslRsa;
+use std::sync::Arc;
 
 #[cfg(all(not(feature = "dnssec-ring"), feature = "dnssec-openssl"))]
 use super::openssl::{Ec, Rsa};
 #[cfg(feature = "dnssec-ring")]
 use super::ring::{Ec, Ed25519, Rsa};
 use super::Algorithm;
-use crate::error::{DnsSecResult, ProtoResult};
+use crate::error::ProtoResult;
 
 /// PublicKeys implement the ability to ideally be zero copy abstractions over public keys for verifying signed content.
 ///
@@ -88,52 +76,6 @@ impl PublicKeyBuf {
     /// Constructs a new PublicKey from the specified bytes, these should be in DNSKEY form.
     pub fn new(key_buf: Vec<u8>, algorithm: Algorithm) -> Self {
         Self { key_buf, algorithm }
-    }
-
-    /// Constructs a new [`PublicKeyBuf`] from an [`OpenSslRsa`] key.
-    #[cfg(feature = "dnssec-openssl")]
-    pub fn from_rsa<T: HasPublic>(key: &OpenSslRsa<T>, algorithm: Algorithm) -> Self {
-        let mut key_buf = Vec::new();
-
-        // this is to get us access to the exponent and the modulus
-        let e = key.e().to_vec();
-        let n = key.n().to_vec();
-
-        if e.len() > 255 {
-            key_buf.push(0);
-            key_buf.push((e.len() >> 8) as u8);
-        }
-
-        key_buf.push(e.len() as u8);
-        key_buf.extend_from_slice(&e);
-        key_buf.extend_from_slice(&n);
-        Self { key_buf, algorithm }
-    }
-
-    /// Constructs a new [`PublicKeyBuf`] from an openssl [`EcKey`].
-    #[cfg(feature = "dnssec-openssl")]
-    pub fn from_ec<T: HasPublic>(ec_key: &EcKey<T>) -> DnsSecResult<Self> {
-        let group = ec_key.group();
-        let algorithm = match group.curve_name() {
-            Some(Nid::X9_62_PRIME256V1) => Algorithm::ECDSAP256SHA256,
-            Some(Nid::SECP384R1) => Algorithm::ECDSAP384SHA384,
-            val => {
-                return Err(format!(
-                    "unsupported curve {val:?} ({:?})",
-                    val.and_then(|nid| nid.long_name().ok())
-                )
-                .into())
-            }
-        };
-
-        let point = ec_key.public_key();
-        let mut key_buf = BigNumContext::new().and_then(|mut ctx| {
-            point.to_bytes(group, PointConversionForm::UNCOMPRESSED, &mut ctx)
-        })?;
-
-        // Remove OpenSSL header byte
-        key_buf.remove(0);
-        Ok(Self { key_buf, algorithm })
     }
 
     /// Extract the inner buffer of public key bytes.
